+++ python/cudf/cudf/utils/utils.py
@@ -7,7 +7,6 @@ import pyarrow as pa
 from numba import njit
 
 import rmm
-from rmm import rmm_config
 
 mask_dtype = np.dtype(np.int8)
 mask_bitsize = mask_dtype.itemsize * 8
@@ -277,25 +276,25 @@ def _set_rmm_config(
         If ``False`` (default), disable pool mode.
     initial_pool_size : int, optional
         If ``use_pool_allocator=True``, sets initial pool size.
-        If ``None``, us
-es 1/2 of total GPU memory.
+        If ``None``, uses 1/2 of total GPU memory.
     enable_logging : bool, optional
         Enable logging (default ``False``).
         Enabling this option will introduce performance overhead.
     """
-    rmm.finalize()
-    rmm_config.use_managed_memory = use_managed_memory
     if use_pool_allocator:
-        rmm_config.use_pool_allocator = use_pool_allocator
         if initial_pool_size is None:
             initial_pool_size = 0  # 0 means 1/2 GPU memory
         elif initial_pool_size == 0:
             initial_pool_size = 1  # Since "0" is semantic value, use 1 byte
         if not isinstance(initial_pool_size, int):
             raise TypeError("initial_pool_size must be an integer")
-        rmm_config.initial_pool_size = initial_pool_size
-    rmm_config.enable_logging = enable_logging
-    rmm.initialize()
+
+    rmm.reinitialize(
+        pool_allocator=use_pool_allocator,
+        managed_memory=use_managed_memory,
+        initial_pool_size=initial_pool_size,
+        logging=enable_logging,
+    )
 
 
 @initfunc